feat: add flip animation between signup and signin modal
