+++ packages/ui/src/components/authmodal/AuthModal.module.css
@@ -1,0 +1,154 @@
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+
+.modalContainer {
+  width: 370px;
+  height: 545px;
+  background-color: white;
+  border-radius: 15px;
+  display: flex;
+  flex-direction: row;
+  position: relative;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
+  transform-style: preserve-3d;
+  transition: transform 0.6s ease-in-out;
+}
+
+
+.flipped {
+  transform: rotateY(180deg);
+}
+
+
+.modalFront,
+.modalBack {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 20px;
+  position: absolute;
+  backface-visibility: hidden;
+  z-index: 0;
+}
+
+
+.modalFront {
+  position: absolute;
+  border-radius: 15px;
+}
+
+
+.modalBack {
+  background-color: #ffffff; 
+  transform: rotateY(180deg);
+  z-index: 1;
+  border-radius: 15px;
+}
+
+
+.logoWrapperFront {
+  position: absolute;
+  top: 5px;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 1;
+}
+
+.logoWrapperBack {
+  position: absolute;
+  top: 50px;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 1;
+}
+
+.logo {
+  width: 140px;
+  height: auto;
+}
+
+
+.closeButtonFront,
+.closeButtonBack {
+  position: absolute;
+  top: 1px;
+  right: 2px;
+  background: none;
+  border: none;
+  font-size: 2rem;
+  cursor: pointer;
+  color: #666;
+  z-index: 2;
+}
+
+.closeButtonFront:hover,
+.closeButtonBack:hover {
+  color: #333;
+}
+
+.horizontalLine {
+  border: none;
+  border-top: 2px solid #ccc; /* Light gray line */
+  margin-bottom: 16px;
+  width: 80%;
+}
+
+
+.toggleBtn {
+  background: none;
+  border: none;
+  color: #535455;
+  cursor: pointer;
+  font-size: 1rem;
+  text-align: center;
+  margin-top: -1rem;
+}
+
+
+
+
+
+@media (max-width: 480px) {
+
+  .modalContainer {
+      width: 80%;
+      max-height: 80%;
+  }
+
+
+  .logoWrapper {
+      top: 5px;
+      width: 80px;
+  }
+
+  .logo {
+      width: 80px;
+  }
+
+
+  .closeButtonFront,
+  .closeButtonBack {
+      font-size: 1.5rem;
+      top: 8px;
+      right: 8px;
+  }
+
+
+  .toggleBtn {
+      font-size: 0.9rem;
+  }
+}
