+++ packages/ui/src/components/common/input/CustomInput.module.css
@@ -1,6 +1,6 @@
 .customInputGroup {
   position: relative;
-  margin: 12px 0;
+  margin: 10px 0;
   width: 100%;
 }
