--- packages/ui/src/components/signup/Signup.module.css
@@ -1,111 +1,0 @@
-.pageDiv {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: auto;
-  max-width: 100%;
-}
-
-.formTitle {
-  color: var(--black);
-  font-size: 2rem;
-  text-align: center;
-}
-
-.formBox {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  max-width: 400px;
-  background: var(--white);
-  border-bottom: var(--border);
-}
-
-.signbutton {
-  margin: 1rem 0;
-}
-
-.inputActiontext {
-  padding-top: 1rem;
-  font-size: 1rem;
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: 3rem;
-  color: var(--description);
-  font-weight: 500;
-  cursor: pointer;
-}
-
-.inputActiontext:hover {
-  color: var(--primary);
-}
-
-.alertmessage {
-  height: 2rem;
-  width: 100%;
-  max-width: 400px;
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-top: 1rem;
-  padding-bottom: 1rem;
-}
-.errorMessage {
-  color: #ff3333;
-  font-size: 0.9rem;
-  width: 100%;
-  max-width: 400px;
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.successMessage {
-  color: green;
-  font-size: 1rem;
-  text-align: center;
-  border-radius: 4px;
-  width: 100%;
-  max-width: 400px;
-  line-height: 1.2rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-@media (max-width: 768px) {
-  .formBox {
-    padding: 1rem;
-    width: 90%;
-    border: none;
-  }
-
-  .input {
-    width: 100%;
-    max-width: 100%;
-  }
-}
-
-@media (max-width: 480px) {
-  .formTitle {
-    font-size: 1.5rem;
-  }
-
-  .formBox {
-    padding: 1rem;
-    width: 90%;
-    max-width: 350px;
-    border: none;
-  }
-
-  .input {
-    width: 100%;
-    max-width: 100%;
-  }
-}
